refactor helpers/utils
