+++ web/site/.gitignore
@@ -139,6 +139,8 @@ out
 
 # playwright test results
 test-results
+# playwright auth user info
+tests/e2e/.auth/**
 
 #firebase service account
 service-account.json
