+++ web/site/tests/e2e/helpers/index.ts
@@ -1,0 +1,4 @@
+export * from './assertCommonPageElements'
+export * from './assertH1Text'
+export * from './assertNuxtContent'
+export * from './assertAlertText'
